+++ src/plugins/e-acsl/tests/bts/oracle/gen_bts1478.c
@@ -86,9 +86,9 @@ void __e_acsl_globals_init(void)
   static char __e_acsl_already_run = 0;
   if (! __e_acsl_already_run) {
     __e_acsl_already_run = 1;
-    __e_acsl_store_block((void *)(& global_i_ptr),(size_t)8);
+    __e_acsl_store_block((void *)(& global_i_ptr),8UL);
     __e_acsl_full_init((void *)(& global_i_ptr));
-    __e_acsl_store_block((void *)(& global_i),(size_t)4);
+    __e_acsl_store_block((void *)(& global_i),4UL);
     __e_acsl_full_init((void *)(& global_i));
   }
   return;
@@ -104,7 +104,7 @@ void __e_acsl_globals_clean(void)
 int main(void)
 {
   int __retres;
-  __e_acsl_memory_init((int *)0,(char ***)0,(size_t)8);
+  __e_acsl_memory_init((int *)0,(char ***)0,8UL);
   __e_acsl_globals_init();
   __gen_e_acsl_loop();
   __retres = 0;
